+++ src/app.html
@@ -24,7 +24,7 @@
 
 		%sveltekit.head%
 	</head>
-	<body data-sveltekit-preload-data="hover" class="bg-white dark:bg-black">
+	<body data-sveltekit-preload-data="hover" class="text-background-on-light dark:text-background-on-dark bg-background-light dark:bg-background-dark">
 		<div style="display: contents">%sveltekit.body%</div>
 	</body>
 </html>
